+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added new resource to manage MX records
 - xDnsServerZoneScope
   - Added integration tests ([issue #177](https://github.com/dsccommunity/xDnsServer/issues/177)).
+  - New read-only property `ZoneFile` was added to return the zone scope
+    file name used for the zone scope.
 
 ### Changed
 
@@ -102,6 +104,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - xDnsServerAdZone
   - Now the parameter `ComputerName` can be used without throwing an exception
     ([issue 79](https://github.com/PowerShell/xDnsServer/issues/79)).
+- xDnsServerZoneScope
+  - Correctly returns the zone scope name when calling `Get-TargetResource`.
 
 ## [1.16.0.0] - 2019-10-30